+++ solvers/euler.cpp
@@ -5,7 +5,7 @@
 #include "euler.h"
 
 EulerSolver::EulerSolver(std::shared_ptr<RhsFunction> rhsFunction, std::function<void(State)> onUpdateConsumer_,
-                         double h_, int maxIterations_) : AbstractSolver(rhsFunction, onUpdateConsumer_, h_,
+                         double h_, long long maxIterations_) : AbstractSolver(rhsFunction, onUpdateConsumer_, h_,
                                                                          maxIterations_) {}
 
 State EulerSolver::step() {
